requirements.txt
pandas
python-dotenv
numpy
streamlit
plotly
hydra-core
torch
chariots
scikit-learn
janome
allennlp